requirements/base.txt
fbs[sentry] @ http://build-system.fman.io/pro/b5aab865-bd29-4f23-992c-0eb4f3a24f33/0.9.4
PyQt5==5.15.11
PyInstaller==6.19.0
rsa==4.9
tinycss==0.4
boto3==1.35.99
requests==2.32.3
PyMuPDF==1.27.2.2
Pillow==11.2.1